requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
modbus-tk==1.1.2
PyYAML==6.0
urllib3==1.26.7
